@@ -1,16 +1,22 @@
 {
   "name": "kaos",
   "description": "Kaos Browser Game",
+  "engines": {
+    "node": ">=20.10.0"
+  },
   "author": "Brian Sutherland <bsuth>",
   "contributors": [
     "Robert Sutherland <0p3r4t0r>"
   ],
-  "version": "0.1.0",
+  "maintainers": [
+    "Robert Sutherland <0p3r4t0r>"
+  ],
+  "version": "0.2.0",
   "license": "GPL-3.0",
-  "homepage": "https://kagi-game.app",
+  "homepage": "https://github.com/0p3r4t0r/kaos",
   "scripts": {
     "build": "webpack --config webpack.config.production.js",
-    "prod": "python3 -m http.server --directory dist",
+    "prod": "npx http-server -o dist",
     "build:dev": "webpack --config webpack.config.development.js",
     "dev": "webpack-dev-server --config webpack.config.development.js",
     "start": "npm run build; npm run prod",
@@ -18,36 +24,34 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/kaos-game/kaos.git"
+    "url": "git+https://github.com/0p3r4t0r/kaos"
   },
   "dependencies": {
-    "is_js": "^0.9.0",
-    "swiper": "^5.4.1",
-    "vue": "^2.6.11",
-    "vue-i18n": "^8.18.1",
-    "vue-router": "^3.1.6",
-    "vue-template-compiler": "^2.6.11",
-    "vuex": "^3.4.0"
+    "swiper": "11.0.05",
+    "vue": "3.3.11",
+    "vue-i18n": "9.8.0",
+    "vue-router": "4.2.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.6",
-    "@babel/preset-env": "^7.9.6",
-    "autoprefixer": "^9.7.6",
-    "babel-loader": "^8.1.0",
-    "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^5.1.1",
-    "css-loader": "^3.5.2",
-    "eslint": "^6.8.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "html-webpack-plugin": "^4.2.0",
-    "mini-css-extract-plugin": "^0.9.0",
+    "@babel/core": "7.23.6",
+    "@babel/preset-env": "7.23.6",
+    "autoprefixer": "10.4.16",
+    "babel-loader": "9.1.3",
+    "clean-webpack-plugin": "4.0.0",
+    "copy-webpack-plugin": "11.0.0",
+    "css-loader": "6.8.1",
+    "eslint": "8.55.0",
+    "eslint-plugin-vue": "9.19.2",
+    "html-webpack-plugin": "5.5.4",
+    "mini-css-extract-plugin": "2.7.6",
     "nipplejs": "^0.8.5",
-    "node-sass": "^4.14.1",
-    "postcss-loader": "^3.0.0",
-    "sass-loader": "^8.0.2",
-    "vue-loader": "^15.9.1",
-    "webpack": "^4.42.1",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.11.0"
+    "node-sass": "9.0.0",
+    "postcss-loader": "7.3.3",
+    "sass-loader": "13.3.2",
+    "vue-loader": "17.3.1",
+    "vue-style-loader": "4.1.3",
+    "webpack": "5.89.0",
+    "webpack-cli": "5.1.4",
+    "webpack-dev-server": "4.15.1"
   }
 }
