@@ -1,44 +1,34 @@
-<!--
- Kaos
- Copyright (C) 2020 Brian Sutherland (bsuth)
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-
 <template>
-    <div>
-        <!-- swiper wrapper -->
-        <div class='swiper-container'>
-            <!-- slide wrapper -->
-            <div class='swiper-wrapper'>
-                <!-- slides -->
-                <div v-for='(slide, index) in slides' :key='index' class='swiper-slide'>
-                    <object :data='slide.data' type='image/svg+xml' />
-                    <div class='divider' />
-                    <h3>{{ $t(slide.label) }}</h3>
-                </div>
-            </div>
-
-            <!-- pagination -->
-            <div class='swiper-pagination'></div>
+  <div>
+    <!-- swiper wrapper -->
+    <div class="swiper-container">
+      <!-- slide wrapper -->
+      <div class="swiper-wrapper">
+        <!-- slides -->
+        <div
+          v-for="(slide, index) in slides"
+          :key="index"
+          class="swiper-slide"
+        >
+          <object
+            :data="slide.data"
+            type="image/svg+xml"
+          />
+          <div class="divider" />
+          <h3>{{ $t(slide.label) }}</h3>
         </div>
+      </div>
+
+      <!-- pagination -->
+      <div class="swiper-pagination" />
     </div>
+  </div>
 </template>
 
 
 <script>
-import Swiper from 'swiper';
-import 'swiper/css/swiper.min.css';
+import Swiper from 'swiper/bundle';
+import 'swiper/css/bundle';
 import { ACTION_EVENTS } from 'input/events';
 
 const INVALID_SWIPER_ACCEPT_CLASSES = [
@@ -48,10 +38,6 @@ const INVALID_SWIPER_ACCEPT_CLASSES = [
 ];
 
 export default {
-    methods: {
-        next: function() { this.swiper.slideNext(); },
-        prev: function() { this.swiper.slidePrev(); },
-    },
 
     data() {
         return {
@@ -81,9 +67,13 @@ export default {
         window.addEventListener(ACTION_EVENTS.LEFT, this.prev);
     },
 
-    beforeDestroy() {
+    beforeUnmount() {
         window.removeEventListener(ACTION_EVENTS.RIGHT, this.next);
         window.removeEventListener(ACTION_EVENTS.LEFT, this.prev);
+    },
+    methods: {
+        next: function() { this.swiper.slideNext(); },
+        prev: function() { this.swiper.slidePrev(); },
     },
 };
 </script>
