@@ -1,24 +1,8 @@
-/*
- * Kaos
- * Copyright (C) 2020 Brian Sutherland (bsuth)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import Vue from 'vue';
+import { createApp } from 'vue';
 
 import App from './App.vue';
 import locale from './locale';
+import routes from './Main/router';
 import 'input';
 import './style/root.scss';
 
@@ -26,8 +10,10 @@ import './style/root.scss';
 // APP
 // -----------------------------------------------------------------------------
 
-new Vue({
-    el: '#app',
-    render: h => h(App),
-    i18n: locale,
-});
+const app = createApp(App);
+
+app.use(locale);
+app.use(routes);
+app.mount('#app');
+
+export default app;
