@@ -1,42 +1,29 @@
-<!--
- Kaos
- Copyright (C) 2020 Brian Sutherland (bsuth)
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-
 <template>
-    <div class='dialog'>
-        <span class='dialog-msg'>
-            <slot />
-        </span>
-        <List
-            :items='items'
-            :activeIndex='activeIndex'
-        />
-    </div>
+  <div class="dialog">
+    <span class="dialog-msg">
+      <slot />
+    </span>
+    <List
+      :items="items"
+      :active-index="activeIndex"
+    />
+  </div>
 </template>
 
 
 <script>
-import { ACTION_EVENTS } from 'input/events';
 import { setContext, CONTEXTS } from 'input/state';
 
 import List from 'components/List.vue';
 
 export default {
     components: { List },
-    props: [ 'items' ],
+    props: {
+        items: {
+            type: Array,
+            required: true,
+        },
+    },
 
     data() {
         return {
@@ -48,7 +35,7 @@ export default {
         setContext(CONTEXTS.MENU);
     },
 
-    beforeDestroy() {
+    beforeUnmount() {
         setContext(CONTEXTS.GAME);
     },
 };
