@@ -1,35 +1,26 @@
-<!--
- Kaos
- Copyright (C) 2020 Brian Sutherland (bsuth)
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-
 <template>
-    <div id='hud'>
-        <!-- {{ $t('hud.nextcolor') }}<div class='player-color' v-bind:class='nextColorClass' /> -->
-        {{ $t('hud.score') }}<span class='score'>{{ state.score }}</span>
-    </div>
+  <div id="hud">
+    <!-- {{ $t('hud.nextcolor') }}<div class='player-color' v-bind:class='nextColorClass' /> -->
+    {{ $t('hud.score') }}<span class="score">{{ gameState.score }}</span>
+  </div>
 </template>
 
 
 <script>
-import { state, player } from 'engine/core';
+import { gameState, player } from 'engine/core';
 
 export default {
+
+    data() {
+        return {
+            gameState: gameState,
+            player: player,
+        };
+    },
+    
     computed: {
-        nextColorClass: function() {
-            switch(player.colorId) {
+        nextColorClass: function () {
+            switch (player.colorId) {
             case 0:
                 return 'purple';
             case 1:
@@ -38,15 +29,10 @@ export default {
                 return 'cyan';
             case 3:
                 return 'red';
+            default:
+                throw Error(`Invalid player color: ${player.colorId}`);
             }
         },
-    },
-
-    data() {
-        return {
-            state: state, 
-            player: player, 
-        };
     },
 };
 </script>
@@ -78,8 +64,19 @@ export default {
     text-align: center;
 }
 
-.red { background: $red; }
-.purple { background: $purple; }
-.green { background: $green; }
-.cyan { background: $cyan; }
+.red {
+    background: $red;
+}
+
+.purple {
+    background: $purple;
+}
+
+.green {
+    background: $green;
+}
+
+.cyan {
+    background: $cyan;
+}
 </style>
