@@ -1,24 +1,10 @@
-<!--
- Kaos
- Copyright (C) 2020 Brian Sutherland (bsuth)
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-
 <template>
-    <div id='fullscreen' v-on:click='toggleFullscreen()'>
-        <Sprite :sprite='sprite' />
-    </div>
+  <div
+    id="fullscreen"
+    @click="toggleFullscreen()"
+  >
+    <Sprite :sprite="sprite" />
+  </div>
 </template>
 
 <script>
@@ -34,12 +20,21 @@ export default {
         };
     },
 
+    mounted() {
+        this.icon = this.$el.children[0];
+        document.addEventListener('fullscreenchange', this.fsChange);
+    },
+
+    beforeUnmount() {
+        document.removeEventListener('fullscreenchange', this.fsChange);
+    },
+
     methods: {
         toggleFullscreen: function() {
             if (this.isFullscreen) {
                 document.exitFullscreen();
             } else {
-                let app = this.$root.$children[0];
+                let app = this.$root;
                 app.$el.requestFullscreen();
             }
         },
@@ -47,15 +42,6 @@ export default {
             this.isFullscreen = !this.isFullscreen;
             this.sprite = (this.isFullscreen) ? 'fs-exit' : 'fs-enter';
         },
-    },
-
-    mounted() {
-        this.icon = this.$el.children[0];
-        document.addEventListener('fullscreenchange', this.fsChange);
-    },
-
-    beforeDestroy() {
-        document.removeEventListener('fullscreenchange', this.fsChange);
     },
 };
 </script>
