@@ -1,31 +1,23 @@
-<!--
- Kaos
- Copyright (C) 2020 Brian Sutherland (bsuth)
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-
 <template>
-    <header>
-        <router-link id='logo' to='/'>
-            <div id='logo-white'>
-                <object data='logo/LogoWhite.svg' type='image/svg+xml' />
-            </div>
-            <div id='logo-color'>
-                <object data="logo/LogoColor.svg" type="image/svg+xml" />
-            </div>
-        </router-link>
-    </header>
+  <header>
+    <router-link
+      id="logo"
+      to="/"
+    >
+      <div id="logo-white">
+        <object
+          data="logo/LogoWhite.svg"
+          type="image/svg+xml"
+        />
+      </div>
+      <div id="logo-color">
+        <object
+          data="logo/LogoColor.svg"
+          type="image/svg+xml"
+        />
+      </div>
+    </router-link>
+  </header>
 </template>
 
 <script>
