@@ -1,31 +1,17 @@
-<!--
- Kaos
- Copyright (C) 2020 Brian Sutherland (bsuth)
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-
 <template>
-    <table>
-        <tr>
-            <td>{{ $t('leaderboard.date') }}</td>
-            <td>{{ $t('leaderboard.score') }}</td>
-        </tr>
-        <tr v-for='(data, index) in scores' :key='index'>
-            <td>{{ data.date }}</td>
-            <td>{{ data.score }}</td>
-        </tr>
-    </table>
+  <table>
+    <tr>
+      <td>{{ $t('leaderboard.date') }}</td>
+      <td>{{ $t('leaderboard.score') }}</td>
+    </tr>
+    <tr
+      v-for="(data, index) in scores"
+      :key="index"
+    >
+      <td>{{ data.date }}</td>
+      <td>{{ data.score }}</td>
+    </tr>
+  </table>
 </template>
 
 
