@@ -1,25 +1,8 @@
-<!--
- Kaos
- Copyright (C) 2020 Brian Sutherland (bsuth)
-
- This program is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program.  If not, see <http://www.gnu.org/licenses/>.
--->
-
 <template>
-    <List
-        :items='items'
-        :activeIndex='activeIndex'
-    />
+  <List
+    :items="items"
+    :active-index="activeIndex"
+  />
 </template>
 
 <script>
@@ -27,16 +10,6 @@ import List from 'components/List';
 
 export default {
     components: { List },
-
-    methods: {
-        goto: function(route) {
-            // VueRouter throws an exception if we try to push the same route
-            // we are already on.
-            if (this.$route.path == route)
-                return;
-            this.$router.push(route);
-        },
-    },
 
     data() {
         return {
@@ -55,7 +28,17 @@ export default {
                     action: () => this.goto('/settings'),
                 },
             ],
-        }
+        };
+    },
+
+    methods: {
+        goto: function(route) {
+            // VueRouter throws an exception if we try to push the same route
+            // we are already on.
+            if (this.$route.path == route)
+                return;
+            this.$router.push(route);
+        },
     },
 };
 </script>
