@@ -1,22 +1,4 @@
-/*
- * Kaos
- * Copyright (C) 2020 Brian Sutherland (bsuth)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import Vue from 'vue';
-import VueI18n from 'vue-i18n';
+import { createI18n } from 'vue-i18n';
 
 import en from './en';
 import ja from './ja';
@@ -25,9 +7,8 @@ import ja from './ja';
 // LOCALIZATION
 // ----------------------------------------------------------------------------
 
-Vue.use(VueI18n);
-
-export default new VueI18n({
-    locale: 'ja',
+export default createI18n({
+    locale: 'en',
+    fallbackLocale: 'ja',
     messages: { en, ja, },
 });
