@@ -1,27 +1,11 @@
-/*
- * Kaos
- * Copyright (C) 2020 Brian Sutherland (bsuth)
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
 const path = require('path');
 
+const webpack = require('webpack');
 const { CleanWebpackPlugin } = require('clean-webpack-plugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-const VueLoaderPlugin = require('vue-loader/lib/plugin');
+const { VueLoaderPlugin } = require('vue-loader');
 
 // -----------------------------------------------------------------------------
 // PRODUCTION BUILD
@@ -57,6 +41,9 @@ module.exports = {
             {
                 test: /\.vue$/,
                 loader: 'vue-loader',
+                options: {
+                    reactivityTransform: true,
+                }
             },
             {
                 test: /\.(s?)css$/,
@@ -87,13 +74,19 @@ module.exports = {
             filename: 'index.html',
             template: 'src/index.html'
         }),
-        new CopyWebpackPlugin([
-            { from: 'assets' }
-        ]),
+        new CopyWebpackPlugin({
+            patterns: [
+                { from: 'assets' }
+            ]
+        }),
         // exclude VueLoaderPlugin() when using MiniCssExtractPlugin()
         new MiniCssExtractPlugin({
             filename: 'style.css'
         }),
         new VueLoaderPlugin(),
+        new webpack.DefinePlugin({
+            __VUE_OPTIONS_API__: true,
+            __VUE_PROD_DEVTOOLS__: false,
+        })
     ],
 };
